requirements.txt
ipykernel>=6.27.1
ipympl>=0.9.3
Jinja2>=3.1.2
jupyter>=1.0.0
matplotlib-inline>=0.1.6
matplotlib>=3.7.2
numpy>=1.25.1
pandas>=2.0.3
plotly>=5.16.1
PyYAML>=6.0.1
scikit-learn>=1.3.2
scipy>=1.11.1
seaborn>=0.12.2
statsmodels>=0.14.1
tqdm>=4.66.1
wfdb==4.1.2
